# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Timewise HRMS API",
    "description": "API documentation for Timewise HRMS leave management system",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "LeaveRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["Annual", "Sick", "Unpaid", "Other"]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["Pending", "Approved", "Rejected"],
            "default": "Pending"
          }
        },
        "required": ["employeeId", "type", "startDate", "endDate"]
      },
      "Employee": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "department": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user"],
            "default": "user"
          }
        },
        "required": ["firstName", "lastName", "email"]
      }
    }
  },
  "paths": {
    "/employees": {
      "get": {
        "summary": "Get all employees",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Employee"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Create a new employee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/employees/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Get employee by ID",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Employee details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "put": {
        "summary": "Update employee",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Employee"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee updated successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      },
      "delete": {
        "summary": "Delete employee",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Employee deleted successfully"
          },
          "404": {
            "description": "Employee not found"
          }
        }
      }
    },
    "/requests": {
      "get": {
        "summary": "Get all leave requests",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["Pending", "Approved", "Rejected"]
            }
          },
          {
            "name": "employeeId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of leave requests",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeaveRequest"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new leave request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Leave request created successfully"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/requests/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "summary": "Get leave request by ID",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Leave request details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveRequest"
                }
              }
            }
          },
          "404": {
            "description": "Leave request not found"
          }
        }
      },
      "put": {
        "summary": "Update leave request",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Leave request updated successfully"
          },
          "404": {
            "description": "Leave request not found"
          }
        }
      },
      "delete": {
        "summary": "Delete leave request",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Leave request deleted successfully"
          },
          "404": {
            "description": "Leave request not found"
          }
        }
      }
    }
  }
}